requirements.txt
aniso8601==7.0.0
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
cloudinary==1.26.0
dj-database-url==0.5.0
Django==3.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-jwt==0.3.1
graphene==2.1.8
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
idna==2.10
Pillow==8.2.0
promise==2.3
psycopg2-binary==2.9.1
PyJWT==1.7.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.6
